+++ MatrixSDK/Crypto/CrossSigning/MXCrossSigning.m
@@ -107,8 +107,8 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
 
 - (MXCrossSigningInfo *)createKeys:(NSDictionary<NSString *,NSData *> *__autoreleasing  _Nonnull *)outPrivateKeys
 {
-    NSString *myUserId = _crypto.mxSession.matrixRestClient.credentials.userId;
-    NSString *myDeviceId = _crypto.mxSession.matrixRestClient.credentials.deviceId;
+    NSString *myUserId = _crypto.mxSession.myUserId;
+    NSString *myDeviceId = _crypto.mxSession.myDeviceId;
 
     MXCrossSigningInfo *crossSigningKeys = [[MXCrossSigningInfo alloc] initWithUserId:myUserId];
 
@@ -175,7 +175,7 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
 {
     NSLog(@"[MXCrossSigning] crossSignDeviceWithDeviceId: %@", deviceId);
           
-    NSString *myUserId = self.crypto.mxSession.myUser.userId;
+    NSString *myUserId = self.crypto.mxSession.myUserId;
     
     dispatch_async(self.crypto.cryptoQueue, ^{
         
@@ -338,7 +338,7 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
         _crypto = crypto;
         _crossSigningTools = [MXCrossSigningTools new];
         
-        NSString *myUserId = _crypto.mxSession.matrixRestClient.credentials.userId;
+        NSString *myUserId = _crypto.mxSession.myUserId;
         _myUserCrossSigningKeys = [_crypto.store crossSigningKeysForUser:myUserId];
         
         [self computeState];
@@ -354,10 +354,10 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
     BOOL canTrustCrossSigningBefore = self.canTrustCrossSigning;
     MXCrossSigningInfo *myUserCrossSigningKeysBefore = self.myUserCrossSigningKeys;
     
-    NSString *myUserId = _crypto.mxSession.matrixRestClient.credentials.userId;
+    NSString *myUserId = _crypto.mxSession.myUserId;
     _myUserCrossSigningKeys = [_crypto.store crossSigningKeysForUser:myUserId];
     
-    NSLog(@"[MXCrossSigning] refreshState for device %@: current state: %@ (%@)", self.crypto.store.deviceId, @(self.state), self.myUserCrossSigningKeys);
+    NSLog(@"[MXCrossSigning] refreshState for device %@: Current state: %@", self.crypto.store.deviceId, @(self.state));
 
     // Refresh user's keys
     [self.crypto.deviceList downloadKeys:@[myUserId] forceDownload:YES success:^(MXUsersDevicesMap<MXDeviceInfo *> *usersDevicesInfoMap, NSDictionary<NSString *,MXCrossSigningInfo *> *crossSigningKeysMap) {
@@ -375,7 +375,7 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
             [self resetTrust];
         }
         
-        NSLog(@"[MXCrossSigning] refreshState for device %@: updated state: %@ (%@)", self.crypto.store.deviceId, @(self.state), self.myUserCrossSigningKeys);
+        NSLog(@"[MXCrossSigning] refreshState for device %@: Updated state: %@", self.crypto.store.deviceId, @(self.state));
         
         if (success)
         {
@@ -399,7 +399,7 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
 {
     BOOL isUserVerified = NO;
 
-    NSString *myUserId = _crypto.mxSession.myUser.userId;
+    NSString *myUserId = _crypto.mxSession.myUserId;
     if ([myUserId isEqualToString:crossSigningKeys.userId])
     {
         // Can we trust the current cross-signing setup?
@@ -515,6 +515,9 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
     }
     
     _state = state;
+    
+    NSLog(@"[MXCrossSigning] myUserCrossSigningKeys: %@", _myUserCrossSigningKeys);
+    NSLog(@"[MXCrossSigning] state: %@", @(_state));
 }
 
 // Recompute cross-signing trust on all users we know
@@ -580,7 +583,7 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
         return NO;
     }
     
-    NSString *myUserId = _crypto.mxSession.myUser.userId;
+    NSString *myUserId = _crypto.mxSession.myUserId;
     
     // Is it signed by a locally trusted device?
     NSDictionary<NSString*, NSString*> *myUserSignatures = myMasterKey.signatures.map[myUserId];
@@ -606,7 +609,9 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
     
     if (!isMasterKeyTrusted)
     {
-        NSLog(@"[MXCrossSigning] isSelfTrusted: NO (MSK not trusted)");
+        NSLog(@"[MXCrossSigning] isSelfTrusted: NO (MSK not trusted). MSK: %@", myMasterKey);
+        NSLog(@"[MXCrossSigning] isSelfTrusted: My user devices: %@", [self.crypto.store devicesForUser:myUserId]);
+
         return NO;
     }
     
@@ -615,7 +620,8 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
     BOOL isUSKSignatureValid = [self checkSignatureOnKey:myUserKey byKey:myMasterKey userId:myUserId];
     if (!isUSKSignatureValid)
     {
-        NSLog(@"[MXCrossSigning] isSelfTrusted: NO (Invalid MSK signature for USK)");
+        NSLog(@"[MXCrossSigning] isSelfTrusted: NO (Invalid MSK signature for USK). USK: %@", myUserKey);
+        NSLog(@"[MXCrossSigning] isSelfTrusted: MSK: %@", myMasterKey);
         return NO;
     }
     
@@ -624,7 +630,8 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
     BOOL isSSKSignatureValid = [self checkSignatureOnKey:mySelfKey byKey:myMasterKey userId:myUserId];
     if (!isSSKSignatureValid)
     {
-        NSLog(@"[MXCrossSigning] isSelfTrusted: NO (Invalid MSK signature for SSK)");
+        NSLog(@"[MXCrossSigning] isSelfTrusted: NO (Invalid MSK signature for SSK). SSK: %@", mySelfKey);
+        NSLog(@"[MXCrossSigning] isSelfTrusted: MSK: %@", myMasterKey);
         return NO;
     }
     
@@ -705,7 +712,7 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
            success:(void (^)(void))success
            failure:(void (^)(NSError *error))failure
 {
-    NSString *myUserId = _crypto.mxSession.myUser.userId;
+    NSString *myUserId = _crypto.mxSession.myUserId;
 
     NSDictionary *object = @{
                              @"algorithms": device.algorithms,
@@ -770,7 +777,7 @@ NSString *const MXCrossSigningErrorDomain = @"org.matrix.sdk.crosssigning";
 {
     [self crossSigningKeyWithKeyType:keyType success:^(NSString *publicKey, OLMPkSigning *signing) {
 
-        NSString *myUserId = self.crypto.mxSession.myUser.userId;
+        NSString *myUserId = self.crypto.mxSession.myUserId;
 
         NSError *error;
         NSDictionary *signedObject = [self.crossSigningTools pkSignObject:object withPkSigning:signing userId:myUserId publicKey:publicKey error:&error];