+++ MatrixSDK/Data/MXEventTimeline.m
@@ -863,7 +863,7 @@ NSString *const kMXRoomInviteStateEventIdPrefix = @"invite-";
     if (_isLiveTimeline && (direction == MXTimelineDirectionForwards))
     {
         // Check for local echo suppression
-        if (room.outgoingMessages.count && [event.sender isEqualToString:room.mxSession.myUser.userId])
+        if (room.outgoingMessages.count && [event.sender isEqualToString:room.mxSession.myUserId])
         {
             MXEvent *localEcho = [room pendingLocalEchoRelatedToEvent:event];
             if (localEcho)